requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and storage
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AI/ML dependencies
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
Pillow==10.1.0
imagehash==4.3.1

# Image processing
opencv-python==4.8.1.78
numpy==1.24.3

# Cryptography
cryptography==41.0.8
pycryptodome==3.19.0

# QR codes and PDFs
qrcode[pil]==7.4.2
reportlab==4.0.7

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: GPU acceleration (uncomment if using CUDA)
# torch==2.1.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision==0.16.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
